NF: Move mathjax method in its own class
